LayoutTests/imported/w3c/web-platform-tests/viewport/viewport-unscaled-scale-iframe-expected.txt

CONSOLE MESSAGE: line 31: TypeError: undefined is not an object (evaluating 'frames[0].window.visualViewport.scale')
Viewport: Unscaled scale iframe

Test Description: This test checks that the default value for scale inside an iframe is 1.


iframe's window.visualViewport.scale is .


Harness Error (FAIL), message = TypeError: undefined is not an object (evaluating 'frames[0].window.visualViewport.scale')

FAIL iframe's visualViewport.scale default value. undefined is not an object (evaluating 'frames[0].window.visualViewport.scale')
